fix(chat): move the extended-thinking toggle next to the model selector (#700)
The thinking toggle lived in the chat panel's top header, away from the
model it relates to, and was easy to miss. Move it into the status bar
right beside the model selector (a small 12px Brain icon, tinted when
on), since thinking capability is a property of the chosen model. Same
gating — only shown for adaptive-thinking models, hidden for Haiku 4.5.

Removed the header toggle and its now-unused imports/state from
ChatPanel; the toggle reads the same chatStore thinkingEnabled flag.

Refs #700
